fix(python): resolve circular import in dora package after adora→dora rename
PR #186's rename created dora/__init__.py with `from dora import ...`
which circularly imports itself instead of the PyO3 native module.

Fix: replace the broken shim with the real package init (from the old
adora/__init__.py) that uses relative imports `from .dora import *`.
Flatten the directory structure — move dora/adora/* up to dora/*.

Also skip rust-dataflow-git example in CI (depends on iceoryx which
needs libacl1-dev).
